change dvbpsi_packet_push to take a const data pointer
dvbpsi_packet_push does not modify the contents of the packet passed to it in
any way : thus, there is no reason for it not to be const.
external code using this function will not be affected by this change.

(cherry picked from commit 0e76e010506b088dc4a83f84086f34be86d00256)