decoder chain: handle unused PSI tables in dvbpsi_chain_demux_delete()
If a dvbpsi_t handle was created but the attached PSI table demuxer was
never triggered because the table is not present in the stream, then the
function dvbpsi_chain_demux_delete() would not free its resources properly.
This caused dvbpsi_delete() on the dvbpsi_t handle to fail.

The function dvbpsi_chain_demux_delete() is changed to handle this situation
properly.