SIS: correctly handle pts_adjustment highest bit (CID #17235)
The code meant to move the upmost pts_adjustment bit 33 down to the
lowest bit (bit 1), so it can be or-ed to a byte as lowest bit. However
the first part of the calculation (i_pts_adjustment & 0x00800) had as result
zero. Then shifting it over 32 bit, resulted in still zero. The calculation
was plain wrong.

First shift all bits over 32 to the right. This places bit 33 on bit 1, then
mask out all the other bits and keep bit 1, thus; (i_pts_adjustment >> 32) & 0x01